Rename Fossil to Cobble
Rename Fossil to Cobble. This will probably block the BLE branch merge.
