+++ android/app/build.gradle
@@ -38,7 +38,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "io.rebble.fossil"
+        applicationId "io.rebble.cobble"
         minSdkVersion 19
         targetSdkVersion 29
         versionCode flutterVersionCode.toInteger()
