+++ android/app/src/main/kotlin/io/rebble/cobble/BlueCommon.kt
@@ -1,4 +1,4 @@
-package io.rebble.fossil
+package io.rebble.cobble
 
 import android.bluetooth.BluetoothDevice
 import android.content.BroadcastReceiver
