+++ android/app/src/main/kotlin/io/rebble/cobble/errors/GlobalErrorHandler.kt
@@ -1,4 +1,4 @@
-package io.rebble.fossil.errors
+package io.rebble.cobble.errors
 
 import android.util.Log
 import kotlinx.coroutines.CoroutineExceptionHandler
@@ -11,7 +11,7 @@ import kotlin.coroutines.CoroutineContext
 class GlobalExceptionHandler @Inject constructor() : CoroutineExceptionHandler {
     override fun handleException(context: CoroutineContext, exception: Throwable) {
         // TODO properly handle exceptions (logging?)
-        Log.e("Fossil", "Coroutine exception", exception)
+        Log.e("Cobble", "Coroutine exception", exception)
     }
 
     override val key: CoroutineContext.Key<*>
