+++ lib/infrastructure/datasources/sqlite/data_transfer_objects/baseobj_dto.dart
@@ -1,8 +1,8 @@
-import 'package:fossil/domain/entities/BaseObj.dart';
-import 'package:fossil/domain/entities/PebbleDevice.dart';
-import 'package:fossil/ui/exceptions/ErrorMessage.dart';
+import 'package:cobble/domain/entities/BaseObj.dart';
+import 'package:cobble/domain/entities/PebbleDevice.dart';
+import 'package:cobble/ui/exceptions/ErrorMessage.dart';
 
-import 'package:fossil/infrastructure/datasources/sqlite/data_transfer_objects/pebbledevice_dto.dart';
+import 'package:cobble/infrastructure/datasources/sqlite/data_transfer_objects/pebbledevice_dto.dart';
 
 extension BaseObjSqlExtensions on BaseObj {
   dynamic get sqlMap {
