+++ lib/ui/devoptions/DevOptionsPage.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:fossil/infrastructure/datasources/DevConnection.dart';
-import 'package:fossil/ui/common/icons/fonts/RebbleIconsStroke.dart';
+import 'package:cobble/infrastructure/datasources/DevConnection.dart';
+import 'package:cobble/ui/common/icons/fonts/RebbleIconsStroke.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class DevOptionsPage extends StatefulWidget {
