+++ lib/ui/setup/MoreSetup.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fossil/ui/common/icons/fonts/RebbleIconsStroke.dart';
+import 'package:cobble/ui/common/icons/fonts/RebbleIconsStroke.dart';
 
 class MoreSetup extends StatefulWidget {
   @override
