+++ lib/ui/setup/boot/RebbleSetupSuccess.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fossil/infrastructure/datasources/WebServices.dart';
+import 'package:cobble/infrastructure/datasources/WebServices.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class RebbleSetupSuccess extends StatefulWidget {
